Add some validation and Email Sending part

+++ Controller/UserManagement.js
@@ -1,9 +1,62 @@
 const mongodb = require("mongodb");
 const db = require("../db");
+const nodemailer=require('nodemailer');
+
 let newIDNo;
 let newID;
 
+
+function generatePassword() {
+  var length = 8,
+    charset =
+      "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789",
+    retVal = "";
+  for (var i = 0, n = charset.length; i < length; ++i) {
+    retVal += charset.charAt(Math.floor(Math.random() * n));
+  }
+  return retVal;
+}
+
+
+function sendEmail(email,pwd){
+  const h1="<h2>Login Details for LMS</h2><hr>"
+  const h2=h1+"<h3>UserName: " + email + "<br>Password: "+pwd +"</h3>"
+
+  var transporter = nodemailer.createTransport({
+    service: 'gmail',
+    auth: {
+      user: 'project2020sliit@gmail.com',
+      pass: 'sliit2020'
+    }
+  });
+  
+  var mailOptions = {
+    from: 'project2020sliit@gmail.com',
+    to: email,
+    subject: 'Login Details',
+    html: h2
+    
+  };
+  
+  transporter.sendMail(mailOptions, function(error, info){
+    if (error) {
+      console.log(error);
+      // return false
+    } else {
+      console.log('Email sent: ' + info.response);
+      // return true
+    }
+  });
+
+}
+
+
+
 exports.GetUserID = (req, res, next) => {
+  if (req.auth === false) {
+    res.status(200).json({ auth: false });
+    return;
+  }
   db.getDb()
     .db()
     .collection("User")
@@ -22,27 +75,19 @@ exports.GetUserID = (req, res, next) => {
 };
 
 exports.AddUser = (req, res, next) => {
-  // if (req.auth === false) {
-  //   res.status(200).json({ auth: false });
-  //   return;
-  // }
+  const emailIDtoSend= req.body.email;
+  let password;
+  if (req.auth === false) {
+    res.status(200).json({ auth: false });
+    return;
+  }
   db.getDb()
     .db()
     .collection("User")
     .findOne({ email: req.body.email })
     .then((resp) => {
       if (!resp) {
-        function generatePassword() {
-          var length = 8,
-            charset =
-              "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789",
-            retVal = "";
-          for (var i = 0, n = charset.length; i < length; ++i) {
-            retVal += charset.charAt(Math.floor(Math.random() * n));
-          }
-          return retVal;
-        }
-        const password = generatePassword();
+        password = generatePassword();
 
         db.getDb()
           .db()
@@ -60,28 +105,34 @@ exports.AddUser = (req, res, next) => {
             faculty: req.body.faculty,
           })
           .then((resp) => {
-            res.status(200).json({ notAdded: false });
+            if (resp.insertedId) {
+              res.status(200).json({ notAdded: false });
+              sendEmail(emailIDtoSend,password)
+
+              
+            }
+            else{
+              res.status(200).json({ error: true });
+            }
+            
           })
           .catch((er) => {
-            console.log(er);
+            res.status(200).json({ error: true });
           });
       } else {
         res.status(200).json({ notAdded: true });
       }
     })
-    .catch(() => {
-      console.log("error");
+    .catch((er) => {
+      res.status(200).json({ error: true });
     });
 };
 
 exports.GetUsers = (req, res, next) => {
-  
-  // if (req.auth === false) {
-  //   console.log("-------------");
-  //   res.status(200).json({ auth: false });
-  //   return;
-  // }
-  
+  if (req.auth === false) {
+    res.status(200).json({ auth: false });
+    return;
+  }
 
   db.getDb()
     .db()
@@ -90,34 +141,50 @@ exports.GetUsers = (req, res, next) => {
     .toArray()
     .then((resp) => {
       if (!resp) {
-        res.status(200).json({ noData: true});
+        res.status(200).json({ noData: true });
       } else {
-        // console.log(resp);
         res.status(200).json(resp);
       }
     })
     .catch(() => {
-      res.status(200).json({ dbError: true});
+      res.status(200).json({ dbError: true });
     });
 };
 
 exports.EditUser = (req, res, next) => {
-  console.log(req.body);
+  if (req.auth === false) {
+    res.status(200).json({ auth: false });
+    return;
+  }
+  if (req.query.id.length !== 24) {
+    res.status(200).json({ fetch: false });
+    return;
+  }
   db.getDb()
     .db()
     .collection("User")
-    // .findOne({ email: req.body.id })
-    .findOne({ _id: new mongodb.ObjectId(req.body.id) })
+    .findOne({ _id: new mongodb.ObjectId(req.query.id) })
     .then((resp) => {
-      // console.log(resp);
-      res.status(200).json(resp);
+      if (!resp) {
+        res.status(200).json({ noData: true });
+      } else {
+        res.status(200).json(resp);
+      }
     })
     .catch(() => {
-      console.log("err");
+      res.status(200).json({ dbError: true });
     });
 };
 
 exports.UpdateUser = (req, res, next) => {
+  if (req.auth === false) {
+    res.status(200).json({ auth: false });
+    return;
+  }
+  if (req.body._id.length !== 24) {
+    res.status(200).json({ fetch: false });
+    return;
+  }
   db.getDb()
     .db()
     .collection("User")
@@ -137,24 +204,38 @@ exports.UpdateUser = (req, res, next) => {
       }
     )
     .then((resp) => {
-      res.status(200).json(resp);
-      // console.log(res);
+      if (resp.modifiedCount === 1) {
+        res.status(200).json({ updated: true });
+      } else {
+        res.status(200).json({ updated: false });
+      }
     })
-    .catch(() => {
-      console.log("error");
+    .catch((er) => {
+      res.status(200).json({ updated: false });
     });
 };
 
 exports.DeleteUser = (req, res, next) => {
+  if (req.auth === false) {
+    res.status(200).json({ auth: false });
+    return;
+  }
+  if (req.body._id.length !== 24) {
+    res.status(200).json({ fetch: false });
+    return;
+  }
   db.getDb()
     .db()
     .collection("User")
     .deleteOne({ _id: new mongodb.ObjectId(req.body._id) })
     .then((resp) => {
-      console.log(resp);
-      res.status(200).json(resp);
+      if (resp.deletedCount === 1) {
+        res.status(200).json({ deleted: true });
+      } else {
+        res.status(200).json({ deleted: false });
+      }
     })
     .catch(() => {
-      console.log("err");
+      res.status(200).json({ deleted: false });
     });
 };
